# resources/docs/blog/openapi/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Swagger/OpenAPI documentation for the backend service.",
        "title": "Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/dashboard/articles": {
            "get": {
                "description": "page through articles in dashboard",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard articles"
                ],
                "summary": "List dashboard articles",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_khanzadimahdi_testproject_application_dashboard_article_getArticles.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "update existing article fields, current user must be author",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard articles"
                ],
                "summary": "Update a dashboard article",
                "parameters": [
                    {
                        "description": "Article update payload",
                        "name": "article",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/updatearticle.Request"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "create a new article in dashboard as current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard articles"
                ],
                "summary": "Create dashboard article",
                "parameters": [
                    {
                        "description": "Article data",
                        "name": "article",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/createarticle.Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/createarticle.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/dashboard/comments": {
            "get": {
                "description": "paginated list of comments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard comments"
                ],
                "summary": "List comments",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_khanzadimahdi_testproject_application_dashboard_comment_getComments.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "modify a comment owned by current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard comments"
                ],
                "summary": "Update comment",
                "parameters": [
                    {
                        "description": "Comment update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/updateComment.Request"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "post a comment as authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard comments"
                ],
                "summary": "Create dashboard comment",
                "parameters": [
                    {
                        "description": "Comment data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_khanzadimahdi_testproject_application_dashboard_comment_createComment.Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/dashboard/comments/user": {
            "get": {
                "description": "paginated list of comments authored by current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard comments"
                ],
                "summary": "List my comments",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/getUserComments.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "admin action to edit a user's comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard comments"
                ],
                "summary": "Update other user's comment",
                "parameters": [
                    {
                        "description": "Comment update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/updateUserComment.Request"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/dashboard/my/bookmarks": {
            "get": {
                "description": "paginated list of bookmarks for authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard bookmarks"
                ],
                "summary": "List my bookmarks",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/getUserBookmarks.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "delete a bookmark owned by current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard bookmarks"
                ],
                "summary": "Delete my bookmark",
                "parameters": [
                    {
                        "description": "Bookmark to delete",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/deleteUserBookmark.Request"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "createarticle.Request": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                },
                "excerpt": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "video": {
                    "type": "string"
                }
            }
        },
        "createarticle.Response": {
            "type": "object",
            "properties": {
                "errors": {
                    "$ref": "#/definitions/domain.ValidationErrors"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "deleteUserBookmark.Request": {
            "type": "object",
            "properties": {
                "object_type": {
                    "type": "string"
                },
                "object_uuid": {
                    "type": "string"
                }
            }
        },
...
        "domain.ValidationErrors": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
...
        "getUserBookmarks.Response": {
            "type": "object",
            "properties": {
                "errors": {
                    "$ref": "#/definitions/domain.ValidationErrors"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/getUserBookmarks.bookmarkResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/getUserBookmarks.pagination"
                }
            }
        },
        "getUserBookmarks.bookmarkResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "object_type": {
                    "type": "string"
                },
                "object_uuid": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "getUserBookmarks.pagination": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "getUserComments.Response": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/getUserComments.commentResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/getUserComments.pagination"
                }
            }
        },
        "getUserComments.authorResponse": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "getUserComments.commentResponse": {
            "type": "object",
            "properties": {
                "approved_at": {
                    "type": "string"
                },
                "author": {
                    "$ref": "#/definitions/getUserComments.authorResponse"
                },
                "body": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "object_type": {
                    "type": "string"
                },
                "object_uuid": {
                    "type": "string"
                },
                "parent_uuid": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "getUserComments.pagination": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "getarticles.author": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "getarticles.pagination": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_khanzadimahdi_testproject_application_dashboard_article_getArticles.Response": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_khanzadimahdi_testproject_application_dashboard_article_getArticles.articleResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/getarticles.pagination"
                }
            }
        },
        "github_com_khanzadimahdi_testproject_application_dashboard_article_getArticles.articleResponse": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/getarticles.author"
                },
                "cover": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "video": {
                    "type": "string"
                }
            }
        },
        "github_com_khanzadimahdi_testproject_application_dashboard_comment_createComment.Request": {
            "type": "object",
            "properties": {
                "approved_at": {
                    "type": "string"
                },
                "author_uuid": {
                    "type": "string"
                },
                "body": {
                    "type": "string"
                },
                "object_type": {
                    "type": "string"
                },
                "object_uuid": {
                    "type": "string"
                },
                "parent_uuid": {
                    "type": "string"
                }
            }
        },
        "github_com_khanzadimahdi_testproject_application_dashboard_comment_getComments.Response": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_khanzadimahdi_testproject_application_dashboard_comment_getComments.commentResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_khanzadimahdi_testproject_application_dashboard_comment_getComments.pagination"
                }
            }
        },
        "github_com_khanzadimahdi_testproject_application_dashboard_comment_getComments.authorResponse": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "github_com_khanzadimahdi_testproject_application_dashboard_comment_getComments.commentResponse": {
            "type": "object",
            "properties": {
                "approved_at": {
                    "type": "string"
                },
                "author": {
                    "$ref": "#/definitions/github_com_khanzadimahdi_testproject_application_dashboard_comment_getComments.authorResponse"
                },
                "body": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "object_type": {
                    "type": "string"
                },
                "object_uuid": {
                    "type": "string"
                },
                "parent_uuid": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "github_com_khanzadimahdi_testproject_application_dashboard_comment_getComments.pagination": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "updateComment.Request": {
            "type": "object",
            "properties": {
                "approved_at": {
                    "type": "string"
                },
                "author_uuid": {
                    "type": "string"
                },
                "body": {
                    "type": "string"
                },
                "object_type": {
                    "type": "string"
                },
                "object_uuid": {
                    "type": "string"
                },
                "parent_uuid": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "updateUserComment.Request": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "updatearticle.Request": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                },
                "excerpt": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "video": {
                    "type": "string"
                }
            }
        },
...
    }
}